Refactor so the different componenets are in separate files.
