+++ client/src/pages/Config/Ships/ShipMap/DeckConfig.tsx
@@ -3,17 +3,7 @@ import {useNetRequest} from "client/src/context/useNetRequest";
 import {useParams} from "react-router-dom";
 import PanZoom from "client/src/components/ui/PanZoom";
 import useMeasure from "client/src/hooks/useMeasure";
-import {
-  createContext,
-  MutableRefObject,
-  ReactNode,
-  useContext,
-  useEffect,
-  useImperativeHandle,
-  useMemo,
-  useRef,
-  useState,
-} from "react";
+import {useEffect, useRef, useState} from "react";
 import Button from "@thorium/ui/Button";
 import UploadWell from "@thorium/ui/UploadWell";
 import {netSend} from "client/src/context/netSend";
@@ -21,24 +11,19 @@ import {useConfirm} from "@thorium/ui/AlertDialog";
 import type {
   DeckEdge as DeckEdgeT,
   DeckNode,
-  EdgeFlag,
   NodeFlag,
 } from "server/src/classes/Plugins/Ship/Deck";
-import {useDrag} from "@use-gesture/react";
-import {autoUpdate, offset, shift, useFloating} from "@floating-ui/react-dom";
-import Input from "@thorium/ui/Input";
-import Checkbox from "@thorium/ui/Checkbox";
-import {Portal} from "@headlessui/react";
-import useOnClickOutside from "client/src/hooks/useClickOutside";
-import {useCallback} from "react";
 import {useDeckNode} from "./DeckNodeContext";
+import {NodeCircle} from "./NodeCircle";
+import {EdgeContextProvider} from "./EdgeContextProvider";
+import {DeckEdges} from "./DeckEdges";
 
 export interface PanStateI {
   x: number;
   y: number;
   scale: number;
 }
-type updateNodeParams =
+export type updateNodeParams =
   | {x: number; y: number}
   | {isRoom: boolean}
   | {name: string}
@@ -46,368 +31,6 @@ type updateNodeParams =
   | {flags: NodeFlag[]};
 const pixelRatio = window.devicePixelRatio;
 
-function NodeCircle({
-  id,
-  x,
-  y,
-  isRoom,
-  radius,
-  name,
-  icon,
-  selected,
-  panState,
-  updateNode,
-  selectNode,
-  deselectNode,
-  removeNode,
-  addingEdges,
-  hasCrossDeckConnection,
-}: DeckNode & {
-  panState: MutableRefObject<PanStateI>;
-  updateNode: (params: updateNodeParams) => void;
-  selectNode: () => void;
-  deselectNode: () => void;
-  removeNode: () => void;
-  selected: boolean;
-  addingEdges: boolean;
-  hasCrossDeckConnection: boolean;
-}) {
-  const {
-    x: floatingX,
-    y: floatingY,
-    reference,
-    floating,
-    strategy,
-    update,
-    refs,
-  } = useFloating({
-    placement: "bottom",
-    middleware: [shift(), offset({mainAxis: 10})],
-  });
-  const [radiusValue, setRadiusValue] = useState(radius);
-  useEffect(() => {
-    setRadiusValue(radius);
-  }, [radius]);
-
-  const renderEdges = useTriggerEdgeRender(id);
-  const bind = useDrag(({down, movement: [mx, my]}) => {
-    if (down) {
-      selectNode();
-    }
-    if (!(refs.reference.current instanceof HTMLDivElement)) return;
-    const newX = x * pixelRatio + mx / panState.current.scale;
-    const newY = y * pixelRatio + my / panState.current.scale;
-    refs.reference.current.style.transform = `translate(${newX}px, ${newY}px)`;
-    if (refs.floating.current) {
-      autoUpdate(refs.reference.current, refs.floating.current, update);
-    }
-    renderEdges(newX, newY);
-    if (!down) {
-      updateNode({
-        x: newX / pixelRatio,
-        y: newY / pixelRatio,
-      });
-    }
-  });
-  useOnClickOutside(refs.floating, () => {
-    if (selected && !addingEdges) {
-      deselectNode();
-    }
-  });
-  const confirm = useConfirm();
-  return (
-    <>
-      <div
-        ref={reference}
-        className={`rounded-full ${
-          selected ? (addingEdges ? "bg-purple-400" : "bg-primary") : "bg-white"
-        } w-2 h-2 absolute -top-1 -left-1 cursor-grab touch-none ${
-          hasCrossDeckConnection ? "ring-1" : ""
-        } ring-white ring-offset-1 ring-offset-black`}
-        onMouseDown={e => e.stopPropagation()}
-        {...bind()}
-        style={{
-          transform: `translate(${x * pixelRatio}px, ${y * pixelRatio}px)`,
-        }}
-      >
-        <div
-          className="absolute rounded-full bg-white/10 pointer-events-none"
-          style={{
-            width: `${radiusValue * 2 * pixelRatio}px`,
-            height: `${radiusValue * 2 * pixelRatio}px`,
-            top: `calc(-${radiusValue * pixelRatio}px + 0.25rem)`,
-            left: `calc(-${radiusValue * pixelRatio}px + 0.25rem)`,
-          }}
-        ></div>
-      </div>
-      <Portal>
-        {selected && !addingEdges && (
-          <div
-            ref={floating}
-            className="rounded min-w-max w-44 p-2 bg-black/60 backdrop-blur shadow-lg z-10 text-white space-y-4"
-            style={{
-              position: strategy,
-              top: floatingY ?? "",
-              left: floatingX ?? "",
-            }}
-            onMouseDown={e => e.stopPropagation()}
-          >
-            <Input
-              label="Name"
-              labelHidden
-              autoFocus
-              placeholder="Name"
-              defaultValue={name}
-              onMouseDown={e => e.stopPropagation()}
-              onChange={e => {
-                updateNode({name: e.target.value});
-              }}
-            />
-            <Checkbox
-              label="Is Room"
-              defaultChecked={isRoom}
-              onChange={e => updateNode({isRoom: e.target.checked})}
-            />
-            <Input
-              label="Radius"
-              type="range"
-              min={0}
-              max={100}
-              defaultValue={radius}
-              onMouseDown={e => e.stopPropagation()}
-              onChange={e => setRadiusValue(e.target.valueAsNumber)}
-              onMouseUp={e =>
-                updateNode({
-                  radius: Number(
-                    (e.target as EventTarget & HTMLInputElement).value
-                  ),
-                })
-              }
-            />
-            <Button
-              className="btn-error btn-sm w-full"
-              onClick={async () => {
-                if (
-                  await confirm({
-                    header: "Delete Node",
-                    body: "Are you sure you want to delete this node?",
-                  })
-                ) {
-                  removeNode();
-                }
-              }}
-            >
-              Delete
-            </Button>
-          </div>
-        )}
-      </Portal>
-    </>
-  );
-}
-
-const EdgeContext = createContext<
-  MutableRefObject<Set<(nodeId: number, x: number, y: number) => void>[]>
->(null!);
-
-const EdgeContextProvider = ({children}: {children: ReactNode}) => {
-  const edgeRenderFunctions = useRef<
-    Set<(nodeId: number, x: number, y: number) => void>[]
-  >([]);
-
-  return (
-    <EdgeContext.Provider value={edgeRenderFunctions}>
-      {children}
-    </EdgeContext.Provider>
-  );
-};
-function useEdgeRerender(
-  from: number,
-  to: number,
-  fn: (nodeId: number, x: number, y: number) => void
-) {
-  const callback = useRef(fn);
-  useEffect(() => {
-    callback.current = fn;
-  }, [fn]);
-  const edgeRenderFunctions = useContext(EdgeContext);
-  useEffect(() => {
-    const rerender = callback.current;
-    if (!edgeRenderFunctions.current[from]) {
-      edgeRenderFunctions.current[from] = new Set();
-    }
-    if (!edgeRenderFunctions.current[to]) {
-      edgeRenderFunctions.current[to] = new Set();
-    }
-    edgeRenderFunctions.current[from].add(rerender);
-    edgeRenderFunctions.current[to].add(rerender);
-    return () => {
-      edgeRenderFunctions.current[from].delete(rerender);
-      edgeRenderFunctions.current[to].delete(rerender);
-    };
-  }, [from, to]);
-}
-function useTriggerEdgeRender(nodeId: number) {
-  const edgeRenderFunctions = useContext(EdgeContext);
-
-  return (x: number, y: number) => {
-    if (edgeRenderFunctions.current[nodeId]) {
-      edgeRenderFunctions.current[nodeId].forEach(f => f(nodeId, x, y));
-    }
-  };
-}
-function DeckEdge({
-  id,
-  from,
-  to,
-  allNodes,
-  removeEdge,
-  updateEdge,
-}: {
-  id: number;
-  from: number;
-  to: number;
-  allNodes: DeckNode[];
-  removeEdge: () => void;
-  updateEdge: (input: {weight: number} | {flags: EdgeFlag[]}) => void;
-}) {
-  const fromNode = allNodes.find(node => node.id === from);
-  const toNode = allNodes.find(node => node.id === to);
-
-  const {
-    edgeState: [selectedEdge, setSelectedEdge],
-  } = useDeckNode();
-
-  const {
-    x: floatingX,
-    y: floatingY,
-    reference,
-    floating,
-    strategy,
-    update,
-    refs,
-  } = useFloating({
-    placement: "bottom",
-    middleware: [shift(), offset({mainAxis: 10})],
-  });
-
-  useEdgeRerender(from, to, (nodeId, x, y) => {
-    if (!fromNode || !toNode) return;
-    const {x: fromX, y: fromY} = fromNode;
-    const {x: toX, y: toY} = toNode;
-    if (refs.reference.current instanceof SVGPathElement) {
-      const ref = refs.reference.current;
-      if (nodeId === from) {
-        ref.setAttribute(
-          "d",
-          `M ${x} ${y} L ${toX * pixelRatio} ${toY * pixelRatio}`
-        );
-      } else if (nodeId === to) {
-        ref.setAttribute(
-          "d",
-          `M ${fromX * pixelRatio} ${fromY * pixelRatio} L ${x} ${y}`
-        );
-      }
-    }
-  });
-  const confirm = useConfirm();
-  if (!fromNode || !toNode) return null;
-  const {x: fromX, y: fromY} = fromNode;
-  const {x: toX, y: toY} = toNode;
-  return (
-    <>
-      <path
-        ref={reference}
-        d={`M ${fromX * pixelRatio} ${fromY * pixelRatio} L ${
-          toX * pixelRatio
-        } ${toY * pixelRatio}`}
-        stroke="white"
-        className={`pointer-events-auto ${
-          selectedEdge === id ? "stroke-yellow-500" : "hover:stroke-primary"
-        }`}
-        onClick={() => setSelectedEdge(id)}
-        strokeWidth={2}
-      />
-      <Portal>
-        {selectedEdge === id && (
-          <div
-            ref={floating}
-            className="rounded min-w-max w-44 p-2 bg-black/60 backdrop-blur shadow-lg z-10 text-white space-y-4"
-            style={{
-              position: strategy,
-              top: floatingY ?? "",
-              left: floatingX ?? "",
-            }}
-            onMouseDown={e => e.stopPropagation()}
-          >
-            <Button
-              className="btn-error btn-sm w-full"
-              onClick={async () => {
-                if (
-                  await confirm({
-                    header: "Delete Node",
-                    body: "Are you sure you want to delete this node?",
-                  })
-                ) {
-                  setSelectedEdge(null);
-                  removeEdge();
-                }
-              }}
-            >
-              Delete
-            </Button>
-          </div>
-        )}
-      </Portal>
-    </>
-  );
-}
-function DeckEdges({
-  deckNodes,
-  deckNodeIds,
-}: {
-  deckNodes: DeckNode[];
-  deckNodeIds: number[];
-}) {
-  const {pluginId, shipId, deckName} = useParams() as {
-    pluginId: string;
-    shipId: string;
-    deckName: string;
-  };
-  const data = useNetRequest("pluginShip", {pluginId, shipId});
-
-  return (
-    <svg className="pointer-events-none absolute inset-0 w-full h-full">
-      {data.deckEdges
-        .filter(
-          edge =>
-            deckNodeIds.includes(edge.from) && deckNodeIds.includes(edge.to)
-        )
-        .map(edge => (
-          <DeckEdge
-            key={edge.id}
-            {...edge}
-            allNodes={deckNodes}
-            updateEdge={(input: {weight: number} | {flags: EdgeFlag[]}) => {
-              netSend("pluginShipDeckUpdateEdge", {
-                pluginId,
-                shipId,
-                edgeId: edge.id,
-                ...input,
-              });
-            }}
-            removeEdge={() => {
-              netSend("pluginShipDeckRemoveEdge", {
-                pluginId,
-                shipId,
-                edgeId: edge.id,
-              });
-            }}
-          />
-        ))}
-    </svg>
-  );
-}
 export function DeckConfig() {
   const {pluginId, shipId, deckName} = useParams() as {
     pluginId: string;
